API_estadistica/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=1.26.0
pymongo>=4.6.0
python-dotenv>=1.0.1
